requirements.txt
Flask==1.0.2
Flask-SQLAlchemy
requests
beautifulsoup4
python-dateutil
